Add helper functions that returns callbacks of given tool  (#79) (#80)
* Add helper functions that returns callbacks of given tool + refactor code (#79)

* check io.write error

* fix lint test failure

* change tool existance logic in updater callback

* - use pdtm api to fetch single tool

* - Create example for callback functions
- Add example to gh workflow

* update example file

---------
