+++ .github/workflows/build-test.yml
@@ -33,3 +33,7 @@ jobs:
       - name: Race Condition Tests
         run: go build -race .
         working-directory: cmd/pdtm/
+      
+      - name: Test Example Code
+        run: go run .
+        working-directory: examples/
