+++ internal/runner/runner.go
@@ -1,20 +1,16 @@
 package runner
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net/http"
 	"os"
 	"os/exec"
-	"runtime"
-	"strings"
 
 	"github.com/projectdiscovery/gologger"
 	"github.com/projectdiscovery/pdtm/pkg"
 	"github.com/projectdiscovery/pdtm/pkg/path"
+	"github.com/projectdiscovery/pdtm/pkg/utils"
 )
 
 // Runner contains the internal logic of the program
@@ -31,7 +27,7 @@ func NewRunner(options *Options) (*Runner, error) {
 
 // Run the instance
 func (r *Runner) Run() error {
-	toolList, err := fetchToolList()
+	toolList, err := utils.FetchToolList()
 
 	// if toolList is not nil save/update the cache
 	// else fetch from cache file
@@ -67,7 +63,7 @@ func (r *Runner) Run() error {
 	gologger.Verbose().Msgf("using path %s", r.options.Path)
 
 	for _, tool := range r.options.Install {
-		if i, ok := contains(toolList, tool); ok {
+		if i, ok := utils.Contains(toolList, tool); ok {
 			if err := pkg.Install(r.options.Path, toolList[i]); err != nil {
 				if err == pkg.ErrIsInstalled {
 					gologger.Info().Msgf("%s: %s", tool, err)
@@ -80,7 +76,7 @@ func (r *Runner) Run() error {
 		}
 	}
 	for _, tool := range r.options.Update {
-		if i, ok := contains(toolList, tool); ok {
+		if i, ok := utils.Contains(toolList, tool); ok {
 			if err := pkg.Update(r.options.Path, toolList[i]); err != nil {
 				if err == pkg.ErrIsUpToDate {
 					gologger.Info().Msgf("%s: %s", tool, err)
@@ -91,7 +87,7 @@ func (r *Runner) Run() error {
 		}
 	}
 	for _, tool := range r.options.Remove {
-		if i, ok := contains(toolList, tool); ok {
+		if i, ok := utils.Contains(toolList, tool); ok {
 			if err := pkg.Remove(r.options.Path, toolList[i]); err != nil {
 				var notFoundError *exec.Error
 				if errors.As(err, &notFoundError) {
@@ -142,99 +138,11 @@ func (r *Runner) ListTools(tools []pkg.Tool) error {
 	var i int
 	for _, tool := range tools {
 		i++
-		installedVersion(i, tool)
+		msg := utils.InstalledVersion(tool, au)
+		fmt.Printf("%d. %s %s\n", i, tool.Name, msg)
 	}
 	return nil
 }
 
-func installedVersion(i int, tool pkg.Tool) string {
-	var msg string
-
-	cmd := exec.Command(tool.Name, "--version")
-
-	var outb bytes.Buffer
-	cmd.Stdout = &outb
-	cmd.Stderr = &outb
-	err := cmd.Run()
-	if err != nil {
-		var notFoundError *exec.Error
-		if errors.As(err, &notFoundError) {
-			osAvailable := isOsAvailable(tool)
-			if osAvailable {
-				msg = au.BrightYellow("(not installed)").String()
-			} else {
-				msg = au.Gray(10, "(not supported)").String()
-			}
-		} else {
-			msg = "version not found"
-		}
-	}
-
-	installedVersion := strings.Split(strings.ToLower(outb.String()), "current version: ")
-	if len(installedVersion) == 2 {
-		installedVersionString := strings.TrimPrefix(strings.TrimSpace(string(installedVersion[1])), "v")
-		if strings.Contains(tool.Version, installedVersionString) {
-			msg = au.Green("(latest) (" + tool.Version + ")").String()
-		} else {
-			msg = au.Red("(outdated) ("+installedVersionString+")").String() + " ➡ " + au.Green("("+tool.Version+")").String()
-		}
-	}
-	fmt.Printf("%d. %s %s\n", i, tool.Name, msg)
-	return msg
-}
-
-const host = "https://api.pdtm.sh"
-
-func fetchToolList() ([]pkg.Tool, error) {
-	tools := make([]pkg.Tool, 0)
-
-	// Get current OS name, architecture, and Go version
-	osName := runtime.GOOS
-	osArch := runtime.GOARCH
-	goVersion := runtime.Version()
-
-	// Create the request URL with query parameters
-	reqURL := host + "/api/v1/tools?os=" + osName + "&arch=" + osArch + "&go_version=" + goVersion
-
-	resp, err := http.Get(reqURL)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode == http.StatusOK {
-		body, err := io.ReadAll(resp.Body)
-		if err != nil {
-			return nil, err
-		}
-		err = json.Unmarshal(body, &tools)
-		if err != nil {
-			return nil, err
-		}
-		return tools, nil
-	}
-	return nil, nil
-}
-
-func contains(s []pkg.Tool, toolName string) (int, bool) {
-	for i, a := range s {
-		if strings.EqualFold(a.Name, toolName) {
-			return i, true
-		}
-	}
-	return -1, false
-}
-
 // Close the runner instance
 func (r *Runner) Close() {}
-
-func isOsAvailable(tool pkg.Tool) bool {
-	osData := path.CheckOS()
-	for asset := range tool.Assets {
-		expectedAssetPrefix := tool.Name + "_" + tool.Version + "_" + osData
-		if strings.Contains(asset, expectedAssetPrefix) {
-			return true
-		}
-	}
-	return false
-}
